requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0
motor~=3.7.1
pydantic~=2.11.7
python-dotenv~=1.1.1
pytest~=8.4.1
pytest-asyncio
pytest-cov
certifi==2025.7.14
pymongo~=4.13.2
starlette~=0.47.2
pydantic-settings~=2.7.0
pydantic_core~=2.33.2
httpx~=0.28.1
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
boto3==1.35.92
jwt==1.3.1
botocore==1.35.92

